Update index.html
Added an Error Message popup if an error is detected, as well as an informative message for users that the form must be completed before it can be submitted. Tested in Chrome with debug console. Throws an error re:  the AJAX in line 124 for 'unexpected identifier' (working on that one), but aside from that, the console is showing no other problems!